+++ api/src/main/java/org/allaymc/api/block/dto/Block.java
@@ -12,6 +12,7 @@ import org.allaymc.api.blockentity.BlockEntity;
 import org.allaymc.api.entity.Entity;
 import org.allaymc.api.item.ItemStack;
 import org.allaymc.api.math.MathUtils;
+import org.allaymc.api.math.location.Location3i;
 import org.allaymc.api.math.position.Position3i;
 import org.allaymc.api.math.position.Position3ic;
 import org.allaymc.api.world.Dimension;
@@ -310,6 +311,15 @@ public class Block {
     @SuppressWarnings("ALL")
     public <T extends BlockEntity> T getBlockEntity() {
         return (T) getDimension().getBlockEntity(position);
+    }
+
+    /**
+     * Creates a new {@link Location3i} from this block's position.
+     *
+     * @return a new location with this block's coordinates and dimension
+     */
+    public Location3i getLocation() {
+        return new Location3i(this.position.x(), this.position.y(), this.position.z(), this.getDimension());
     }
 
     @Override