+++ api/src/main/java/org/allaymc/api/block/interfaces/BlockBedBehavior.java
@@ -2,7 +2,10 @@ package org.allaymc.api.block.interfaces;
 
 import org.allaymc.api.block.BlockBehavior;
 import org.allaymc.api.block.component.BlockBlockEntityHolderComponent;
+import org.allaymc.api.block.component.BlockRespawnPointComponent;
 import org.allaymc.api.blockentity.interfaces.BlockEntityBed;
 
-public interface BlockBedBehavior extends BlockBehavior, BlockBlockEntityHolderComponent<BlockEntityBed> {
+public interface BlockBedBehavior extends BlockBehavior,
+        BlockBlockEntityHolderComponent<BlockEntityBed>,
+        BlockRespawnPointComponent {
 }