+++ api/src/main/java/org/allaymc/api/block/interfaces/BlockRespawnAnchorBehavior.java
@@ -1,6 +1,7 @@
 package org.allaymc.api.block.interfaces;
 
 import org.allaymc.api.block.BlockBehavior;
+import org.allaymc.api.block.component.BlockRespawnPointComponent;
 
-public interface BlockRespawnAnchorBehavior extends BlockBehavior {
+public interface BlockRespawnAnchorBehavior extends BlockBehavior, BlockRespawnPointComponent {
 }