+++ api/src/main/java/org/allaymc/api/world/sound/RespawnAnchorChargeSound.java
@@ -1,0 +1,9 @@
+package org.allaymc.api.world.sound;
+
+import org.allaymc.api.block.type.BlockState;
+
+/**
+ * @author IWareQ
+ */
+public record RespawnAnchorChargeSound(BlockState blockState) implements Sound {
+}